+++ app/src/main/java/org/breezyweather/sources/atmoaura/AtmoAuraService.kt
@@ -21,8 +21,8 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import io.reactivex.rxjava3.core.Observable
 import org.breezyweather.BuildConfig
 import org.breezyweather.R
-import org.breezyweather.common.basic.models.Location
-import org.breezyweather.common.basic.wrappers.SecondaryWeatherWrapper
+import breezyweather.domain.location.model.Location
+import breezyweather.domain.weather.wrappers.SecondaryWeatherWrapper
 import org.breezyweather.common.exceptions.ApiKeyMissingException
 import org.breezyweather.common.extensions.getFormattedDate
 import org.breezyweather.common.extensions.toCalendarWithTimeZone
@@ -35,7 +35,9 @@ import org.breezyweather.common.source.SecondaryWeatherSource
 import org.breezyweather.common.source.SecondaryWeatherSourceFeature
 import org.breezyweather.settings.SourceConfigStore
 import retrofit2.Retrofit
-import java.util.*
+import java.util.Calendar
+import java.util.Date
+import java.util.Locale
 import javax.inject.Inject
 
 /**
@@ -94,8 +96,8 @@ class AtmoAuraService @Inject constructor(
 
         return mApi.getPointDetails(
             getApiKeyOrDefault(),
-            location.longitude.toDouble(),
-            location.latitude.toDouble(),  // Tomorrow because it gives access to D-1 and D+1
+            location.longitude,
+            location.latitude, // Tomorrow because it gives access to D-1 and D+1
             calendar.time.getFormattedDate(location.timeZone, "yyyy-MM-dd'T'HH:mm:ss.SSSXXX", Locale.ENGLISH)
         ).map {
             convert(it)
