+++ app/src/main/java/org/breezyweather/sources/metno/json/MetNoMoonProperties.kt
@@ -22,5 +22,5 @@ import kotlinx.serialization.Serializable
 data class MetNoMoonProperties(
     val moonrise: MetNoEphemerisProperty?,
     val moonset: MetNoEphemerisProperty?,
-    val moonphase: Float?,
+    val moonphase: Double?,
 )
