+++ app/src/main/java/org/breezyweather/sources/mf/json/MfHistoryTemperature.kt
@@ -21,6 +21,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MfHistoryTemperature(
-    val value: Float?,
-    @SerialName("windchill") val windChill: Float?
+    val value: Double?,
+    @SerialName("windchill") val windChill: Double?
 )
