+++ app/src/main/java/org/breezyweather/sources/openweather/json/OpenWeatherAirPollutionComponents.kt
@@ -20,12 +20,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class OpenWeatherAirPollutionComponents(
-    val co: Float?,
-    val no: Float?,
-    val no2: Float?,
-    val o3: Float?,
-    val so2: Float?,
-    val pm2_5: Float?,
-    val pm10: Float?,
-    val nh3: Float?
+    val co: Double?,
+    val no: Double?,
+    val no2: Double?,
+    val o3: Double?,
+    val so2: Double?,
+    val pm2_5: Double?,
+    val pm10: Double?,
+    val nh3: Double?
 )
